+++ wiktra/__init__.py
@@ -2,5 +2,5 @@
 # -*- coding: utf-8 -*-
 from .Wiktra import *
 
-__version__ = "2.0.2"
+__version__ = "2.0.3"
 __all__ = ['translite']